requirements.txt
pycryptodome
secretstorage
dbus-python
